droidboot: Use recovery minui for graphics
Droidboot now uses the framebuffer minui library to show some graphics
on the screen. Previously, on most MFLD devices the only visible
output was on the serial console.

The ui.c code and everything under res/ have been copied from
bootable/recovery. The UI code has been stripped down to remove
the menu and keyboard listener. The graphics have been slightly
changed so that the robot shows "DROIDBOOT" in text above it, to
ensure nobody confuses this with recovery mode. The robot itself
is not modified.

The Android robot logo is covered under the Creative Commons Attribution
License and is modifiable under those terms. See
http://www.android.com/branding.html

The UI flow is as follows:
1. Droidboot starts with the robot gears spinning
2. During the execution of any command, a animated indeterminate
progress bar is shown below the robot.
3. If any errors occur (detected by the use of LOGE), the animations
are frozen and the text output is displayed.

Users can force text output display with "fastboot oem showtext".

The entire UI can be disabled by defining DROIDBOOT_NO_GUI := true
in the boardconfig.

The Makefile output in ramdisk.mk has been cleaned up a bit
and made less noisy. The 'droidboot' phony target was renamed
as it was colliding with the module name of the droidboot server.

Change-Id: I84de4f706bafd3f3ef55fccb20cb8ebe1f820a28