add device provisioning mode, misc cleanups
() Droidboot now checks for a file named $TARGET_PRODUCT.auto-ota.zip
on the SD card. If it is found and the user does not cancel the countdown
by pressing a button, then the /data and /cache partitions will be
formatted and the device will reboot into the recovery console to
provision the device with the update file.

() droidboot tag changed to 'eng' so a copy doesn't end up on the
system partition in user builds

() input_listener_thread drastically simplified to use the input event
retrieval code in libminui

() fixed ui_show_text() function prototypes

() use strtoul in fastboot.c since we're not in an LK environment

() move scratch buffer memory allocation to Fastboot code since it's
the only thing that uses it

() bring in fstab reading code from recovery console so that we can
parse the recovery.fstab and determine the device node for the SD card.
Code added to the ramdisk makefile to pull in the recovery.fstab.

() fixed a problem in ui_print() where the bottom row had stale data
or was cut off

() moved some logic out of aboot.c into util.c so it can be used by
other parts of the program

() Use recovery.fstab to determine volume type, and propogate any
filesystem size options to make_ext4fs. This allows compatibility with
systems that use encrypted filesystems, as they need 16K reserved space
at the end of the partition.

Change-Id: Ie8e6ade544a073f54cd87b48061c91f9da125bd1
Depends-Change-Id: Ie7f79b829d2d314798b087b3c0496e1a76c2b5d6