add /cache provisioning capability, cleanups
Building on the code for sdcard provisioning, it is now also possible
to provision the device if an auto-ota file is found in the cache
filesystem. In this case, instead of formatting /cache, it is sized
to fit the partition and some integrity checks are performed on it.

This is used on Medfield by a new type of FSTK image which has a
cache partition containing an OTA update stitched directly into the
image, removing the need for Fastboot commands during provisioning.

A few other adjustments were made:

() Update packages sent over Fastboot get put on cache partition,
not userdata

() Code cleanups in droidboot.c

() Added a plug-in API to register platform-specific actions
taken when provisioning a device for automatic SW updates

() Remove EV_REL from the stimuli which can cancel automatic
actions; we just want buttons

() fstab and util headers renamed; they are intended to be useable
by plug-ins

() Rename LOG* APIs, we're not actually using Android liblog

Change-Id: I9c8444b1a36ca3b1c33ca56bf28489c5d7c95446
Depends-Change-Id: Ie7f79b829d2d314798b087b3c0496e1a76c2b5d6