add runtime config options
The scratch buffer size and autoboot/kexec behavior used to be
configurable with #define, which is not very flexible. Using
AOSP init as a model, we can now tune this stuff at runtime using
droidboot.* parameters stuffed in the kernel command line.

In addition to specifying a filesystem for the 2ndstageboot images,
a containing directory is also specified. We default to
/data/2ndstageboot.

Change-Id: I260d63ba8cfac59f01cb4e3aba4f700c5de9e495
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>