Abort on any fsck errors
The filesystem images made by the build lacked lost+found, which
raised flags with e2fsck. This has been fixed, so make the checking
much more strict; any problems will result in an error without
attempting to fix them and are due to either a malformed filesystem
or data corruption.

Change-Id: I10b82f17f0896134f2fa7c43c572e532276df371
Depends-Change-Id: If9069910c5d3dacc01ae913b870f631bbdab790d
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>