util: Change execute_command() api to use stdarg
() execute_command() changed to accept variable arguments to
make the API much simpler. It is declared with __attribute__((format))
which, combined with -Werror, removes the possibility of several
forms of abuse of this function (non literal format string,
arguments that don't correspond to the format string, etc).
This makes this function as safe to use as regular
printf(), at least in the Droidboot build.

() execute_command_data() which works like execute_command()
but popen()s the target process and feeds it a blob of data
to its stdin.

Change-Id: I9e22b91c61dc39aeb02b724c1d6ad00b0c1a50f8