droidboot: honor the 'length' attribute in recovery.fstab
A 'length' attribute can be specified in recovery.fstab to force
a certain size for a partition or to reserve space at the end of
it (the latter case is required by Android's encryption process
as it needs 16kb of free space at the end of /data partition).
The function ext4_filesystem_checks (util.c) always resized a
partition to its maximum possible size. With this patch,
ext4_filesystem_checks takes into account the 'length' attribute
of a partition, as specified in recovery.fstab.

Issue: ANDROID-673
Change-Id: I03f8143703d8919aab1a814ac780d51594cecd61
Signed-off-by: Bogdan Marinescu <bogdan.a.marinescu@intel.com>