fix filesystem resize
The resize2fs call was complaining that we need to run fsck,
even though we just did with no errors. Pass -f to it to make
it happy.

Return errors on any nonzero return values, not just negative
ones (which was what allowed this bug to go on for so long)

Issue: ANDROID-1583
Change-Id: I80143c48c37135f36ac4d91dff164987773a8cc5
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>