Low memory in /cache after a fastboot flash system
BZ: 51657

Fastboot flash system copies in /cache a file named fastboot.tmp which is not deleted afterwards.
The consequence is a low memory that prevents to push an ota file.
This issue is seen because the flash.xml file does no more perform the flash update that previously cleaned the cache.
Signed-off-by: David Castelain <davidx.castelain@intel.com>

Change-Id: I0551695b2d222ad1547edd2670e045854ec5ca4f
Reviewed-on: http://android.intel.com:8080/61736
Reviewed-by: Castelain, DavidX <davidx.castelain@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Afantenos, Marie CecileX <marie.cecilex.afantenos@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>