Add automatic partitioning and trigger mount in droidboot.
BZ: 89049

Per customer request, droidboot is able to automatically partition the
eMMC IF detected partitions are different from those listed into
recovery.fstab (including blank eMMC) AND IF "droidboot.autopart=1"
is appended to the kernel command line.

Partitions are mounted by default thanks to the property
"sys.partitioning=0" that is set.
But by appending "droidboot.automount=0" to the kernel command line,
the partitions are not mounted.

Change-Id: I410e7f2f6aa68abff35e7887b7c587cec76c8929
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: http://android.intel.com:8080/94346
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Castelain, DavidX <davidx.castelain@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Afantenos, Marie Cecile <marie.cecile.afantenos@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>