Revert volumeutils/roots code to Android default.
BZ: 89060

This code was modified to support MBR partitioning, but had
to get back to standard method avoiding the usage of an Intel
customized recovery.fstab.
MBR is still supported and partitions' size extracted from
a comment in the recovery.fstab. For that, a new function
'get_paritition_size()' was created into ufdisk.c. That
function is largely based on 'load_volume_table()' in roots.c.

Change-Id: I156b507ea1266a25a34ba59b3936905471060178
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: http://android.intel.com:8080/95623
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>