Revert volumeutils/roots code to Android default.
BZ: 89060

(patch re-submitted as a first re-integration step to avoid FOTA issue)
This code was modified to support MBR partitioning, but had
to get back to standard method avoiding the usage of an Intel
customized recovery.fstab.
MBR is still supported and partitions' size extracted from
a comment in the recovery.fstab. For that, a new function
'get_paritition_size()' was created into ufdisk.c. That
function is largely based on 'load_volume_table()' in roots.c.

Change-Id: Iada6667fad32e575ebc26cef9e4a2fc98ab01a29
Signed-off-by: Florent Auger <florent.auger@intel.com>
Reviewed-on: http://android.intel.com:8080/97114
Reviewed-by: Jauffres, MichelX <michelx.jauffres@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Tested-by: Molero Rodenas, BeatrizX <beatrizx.molero.rodenas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>