Revert "Revert volumeutils/roots code to Android default."
BZ: 89060

This code was modified to support MBR partitioning, but had
to get back to standard method avoiding the usage of an Intel
customized recovery.fstab.
MBR is still supported and partitions' size extracted from
a comment in the recovery.fstab. For that, a new function
'get_paritition_size()' was created into ufdisk.c. That
function is largely based on 'load_volume_table()' in roots.c.

Orig-Change-Id: I156b507ea1266a25a34ba59b3936905471060178

This reverts commit 0b22e0790ace88dfed652f1241ada64cd461ca7a.

Change-Id: Iabe38dc751d128fef40bccdb0e83bd567be105ff
Reviewed-on: http://android.intel.com:8080/96807
Reviewed-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>
Tested-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>