Flashing over Ethernet broken.
BZ: 127517

The file descriptor was not initialized when tcp fastboot command occurs.

Initialize usb.read_fp and usb.write_fp as those are globals for fastboot_command_loop().
Change pr_error to pr_info on USB error which is normal if no USB is connected, to avoid too many messages.

Change-Id: I05613c39add665fcb07b6f5c6582da3bfd90b3d8
Signed-off-by: Philippe Lumbierres <philippex.lumbierres@intel.com>
Reviewed-on: http://android.intel.com:8080/122629
Reviewed-by: Muller, Francois-nicolas <francois-nicolas.muller@intel.com>
Reviewed-by: Tinelli, Vincent <vincent.tinelli@intel.com>
Reviewed-by: Betous, GuillaumeX <guillaumex.betous@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Lefevre, Anne-sophie <anne-sophie.lefevre@intel.com>