In droidboot ensure that usb settings are stopped after charging
BZ: 144473

When battery level is less than the minimum level (10 %) the property
sys.usb.config is set to "charging" in droidboot.init.rc "boot" section.
And charging starts.
Once charge level is enough, the droidboot UI will appear and it will be
possible to continue flashing.

Once charged enough, before setting sys.usb.config properties to fastboot,
we need to be sure that previous settings set when charging are unset.

Change-Id: Ica66a3c57c20d503dd45cb68aa00b7eacd41584a
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>