Clean protoype function
BZ: 178474

With ICC the prototype with no arguments is interpreted as function
with no parameter although with GNU compiler, it is interpeted as a
function with unspecified number of parameters.
This leads to an error on ICC compilation because some functions
were not called with the right number of arguments.

Change-Id: I57a6550fe4f86176a12c4ac63127211f3f10f978
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>